soda_check/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.11.1
aws-psycopg2==1.3.8
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
googleapis-common-protos==1.65.0
idna==3.10
importlib-metadata==6.11.0
inflect==7.4.0
Jinja2==3.1.4
MarkupSafe==2.1.2
more-itertools==10.5.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
protobuf==4.25.5
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
soda-core==3.3.22
soda-core-postgres==3.3.22
sqlparse==0.5.1
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
zipp==3.20.2
